requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Faker==8.6.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
